Move host discovery to Discovery struct
Allows other auditing/scanning work to reuse this functionality
